requirements.txt
streamlit>=1.30.0
opencv-python-headless>=4.8.0
openai>=1.12.0
Pillow>=10.0.0
openai-whisper>=20231117
ffmpeg-python>=0.2.0
pandas>=2.0.0
